refactor: 코드 포맷 개선 및 중복 제거
